Crea utils.py para popup msg. Mueve logica add/del cords/cota a armadura_activa.py (Aun no funcinoa abst)
